// src/modules/products/dto/product-data.dto.ts
export class ProductData {
    product_id?: number;
    product_name?: string;
    product_description?: string;
    category_id?: number;
    weight_class?: number;
    warranty_period?: string;
    supplier_id?: number;
    product_status?: string;
    list_price?: number;
    min_price?: number;
    catalog_url?: string;
}